+++ test/Lombiq.OSOCE.Tests.UI/Tests/BehaviorVueTests.cs
@@ -2,7 +2,6 @@
 using Lombiq.Tests.UI.Extensions;
 using Lombiq.Tests.UI.Services;
 using Lombiq.VueJs.Tests.UI.Extensions;
-using OpenQA.Selenium;
 using System.Threading.Tasks;
 using Xunit;
 using Xunit.Abstractions;
@@ -23,25 +22,6 @@ public class BehaviorVueTests : UITestBase
             {
                 await context.SignInDirectlyAsync();
                 await context.TestVueSampleBehaviorAsync();
-            },
-            browser);
-
-    [Theory, Chrome]
-    public Task ConsentBannerShouldWork(Browser browser) =>
-        ExecuteTestAfterSetupAsync(
-            async context =>
-            {
-                var privacyConsentAcceptButton = By.Id("privacy-consent-accept-button");
-
-                // Enable consent banner and interact with it.
-                await context.EnableFeatureDirectlyAsync("Lombiq.Privacy.ConsentBanner");
-                await context.GoToHomePageAsync();
-                await context.ClickReliablyOnAsync(privacyConsentAcceptButton);
-                context.Missing(privacyConsentAcceptButton);
-
-                // Verify persistence.
-                context.Refresh();
-                context.Missing(privacyConsentAcceptButton);
             },
             browser);
 }
